Moved classes in proper packages
