+++ mobile/src/main/java/com/wedriveu/mobile/util/rabbitmq/RabbitMQJsonMapper.java
@@ -1,4 +1,4 @@
-package com.wedriveu.mobile.util;
+package com.wedriveu.mobile.util.rabbitmq;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
